Permissions ProblemeController

+++ src/Controller/ProblemeController.php
@@ -66,63 +66,69 @@ class ProblemeController extends AbstractController
         int $page = 1
     ): Response
     {
-        if(!$this->permissionChecker->isUserGranted(['GET_OTHER_PROBLEME'])){
-            $this->addFlash('fail','vous ne possedez pas les permissions necessaires.');
-            return $this->redirectToRoute('home_index');
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["GET_OTHER_PROBLEME"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            } else {
+                $form = $this->createForm(ProblemeSearchType::class);
+                $form->handleRequest($request);
+
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $tab = $request->request->get("probleme_search");
+
+                    if (array_key_exists("nom", $tab))
+                        $session->set("search_nom_probleme", $tab["nom"]);
+                    else
+                        $session->remove("search_nom_probleme");
+
+                    if (array_key_exists("categories", $tab))
+                        $session->set("search_categories", $tab["categories"]);
+                    else
+                        $session->remove("search_categories");
+
+                    if (array_key_exists("statuts", $tab))
+                        $session->set("search_statuts", $tab["statuts"]);
+                    else
+                        $session->remove("search_statuts");
+
+                    if (array_key_exists("element", $tab))
+                        $session->set("search_element", $tab["element"]);
+                    else
+                        $session->remove("search_element");
+
+                    if (array_key_exists("orderby", $tab))
+                        $session->set("search_orderby", $tab["orderby"]);
+                    else
+                        $session->remove("search_orderby");
+                }
+
+                $active_nom = $session->get('search_nom_probleme') ? $session->get('search_nom_probleme') : "";
+                $active_categories = $session->get('search_categories') ? $session->get('search_categories') : [];
+                $active_statuts = $session->get('search_statuts') ? $session->get('search_statuts') : [];
+                $active_element = $session->get('search_element') ? $session->get('search_element') : 20;
+                $active_orderby = $session->get('search_orderby') ? $session->get('search_orderby') : "priorite";
+
+                $problemes = $problemeRepository->findPaginateByCategoryAndName($page, $active_element, $active_categories, $active_statuts, $active_nom, $active_orderby);
+                $nbr_page = ceil(count($problemeRepository->findAllByCategoryAndName($page, $active_element, $active_categories, $active_statuts, $active_nom)) / $active_element);
+
+                return $this->render('probleme/index.html.twig', [
+                    'problemes' => $problemes,
+                    'nbr_page' => $nbr_page,
+                    'active_page' => $page,
+                    'categories' => $categorieRepository->findAll(),
+                    'active_nom' => $active_nom,
+                    'active_categories' => $active_categories,
+                    'active_statuts' => $active_statuts,
+                    'active_element' => $active_element,
+                    'active_orderby' => $active_orderby,
+                    'form' => $form->createView(),
+                ]);
+            }
         }
-        $form = $this->createForm(ProblemeSearchType::class);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $tab = $request->request->get("probleme_search");
-
-            if(array_key_exists("nom", $tab))
-                $session->set("search_nom_probleme", $tab["nom"]);
-            else
-                $session->remove("search_nom_probleme");
-
-            if(array_key_exists("categories", $tab))
-                $session->set("search_categories", $tab["categories"]);
-            else 
-                $session->remove("search_categories");
-
-            if(array_key_exists("statuts", $tab))
-                $session->set("search_statuts", $tab["statuts"]);
-            else 
-                $session->remove("search_statuts");
-
-            if(array_key_exists("element", $tab))
-                $session->set("search_element", $tab["element"]);
-            else
-                $session->remove("search_element");
-
-            if(array_key_exists("orderby", $tab))
-                $session->set("search_orderby", $tab["orderby"]);
-            else
-                $session->remove("search_orderby");
-        }
-
-        $active_nom = $session->get('search_nom_probleme') ? $session->get('search_nom_probleme') : "";
-        $active_categories = $session->get('search_categories') ? $session->get('search_categories') : [];
-        $active_statuts = $session->get('search_statuts') ? $session->get('search_statuts') : [];
-        $active_element = $session->get('search_element') ? $session->get('search_element') : 20;
-        $active_orderby = $session->get('search_orderby') ? $session->get('search_orderby') : "priorite";
-
-        $problemes = $problemeRepository->findPaginateByCategoryAndName($page, $active_element, $active_categories, $active_statuts, $active_nom, $active_orderby);
-        $nbr_page = ceil(count($problemeRepository->findAllByCategoryAndName($page, $active_element, $active_categories, $active_statuts, $active_nom))/$active_element);
-
-        return $this->render('probleme/index.html.twig', [
-            'problemes' => $problemes,
-            'nbr_page' => $nbr_page,
-            'active_page' => $page,
-            'categories' => $categorieRepository->findAll(),
-            'active_nom' => $active_nom,
-            'active_categories' => $active_categories,
-            'active_statuts' => $active_statuts,
-            'active_element' => $active_element,
-            'active_orderby' => $active_orderby,
-            'form' => $form->createView(),
-        ]);
     }
 
     /**
@@ -228,15 +234,15 @@ class ProblemeController extends AbstractController
                 'probleme' => $probleme,
                 'images' => $imageRepository->findbyProbleme($probleme)
             ]);
-        }else{
-            $canValidate = $this->permissionChecker->isUserGranted(["VALIDATE_PROBLEME"]);
-            $canUpdateStatut = $this->permissionChecker->isUserGranted(["CAN_EDIT_STATUT_PROBLEME"]);
+        }else {
+            $canValidate = $this->permissionChecker->isUserGranted(["UPDATE_OTHER_HISTORIQUE_STATUT"]);
+            $canUpdateStatut = $this->permissionChecker->isUserGranted(["UPDATE_OTHER_HISTORIQUE_STATUT"]);
             return $this->render('probleme/show.html.twig', [
                 'probleme' => $probleme,
                 'images' => $imageRepository->findbyProbleme($probleme),
                 'canValidate' => $canValidate,
                 'canUpdateStatut' => $canUpdateStatut
-                ]);
+            ]);
         }
     }
 
@@ -251,36 +257,42 @@ class ProblemeController extends AbstractController
 
     ): Response
     {
-        if(!$this->permissionChecker->isUserGranted(['UPDATE_OTHER_PROBLEME'])){
-            $this->addFlash('fail','Vous ne possedez pas les permissions necessaires pour modifier ce problème');
-            return $this->redirectToRoute('home_index');
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["UPDATE_OTHER_PROBLEME"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            } else {
+                $images = $imageRepository->findbyProbleme($probleme);
+
+
+                $form = $this->createForm(ProblemeType::class, $probleme);
+                $form->handleRequest($request);
+                $tabImageToProblemes = [];
+                for ($i = 1; $i <= 4; $i++) {
+                    $imageToProbleme = $form['Image' . $i]->getData();
+                    array_push($tabImageToProblemes, $imageToProbleme);
+                }
+                if ($request->query->get('statut')) $render = 'probleme/edit.modal.html.twig';
+                else $render = 'probleme/edit.html.twig';
+
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $problemeService->UploadImagesNewProbleme($tabImageToProblemes, $probleme);
+                    $this->getDoctrine()->getManager()->flush();
+
+                    return $this->redirectToRoute('probleme_show', ['id' => $probleme->getId()]);
+                }
+
+                return $this->render($render, [
+                    'probleme' => $probleme,
+                    'adresse' => $probleme->getLocalisation(),
+                    'form' => $form->createView(),
+                    'images' => $images
+                ]);
+            }
         }
-        $images = $imageRepository->findbyProbleme($probleme);
-
-
-        $form = $this->createForm(ProblemeType::class, $probleme);
-        $form->handleRequest($request);
-        $tabImageToProblemes = [];
-        for ($i = 1; $i <= 4; $i++) {
-            $imageToProbleme = $form['Image' . $i]->getData();
-            array_push($tabImageToProblemes, $imageToProbleme);
-        }
-        if($request->query->get('statut')) $render = 'probleme/edit.modal.html.twig';
-        else $render = 'probleme/edit.html.twig';
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $problemeService->UploadImagesNewProbleme($tabImageToProblemes, $probleme);
-            $this->getDoctrine()->getManager()->flush();
-
-            return $this->redirectToRoute('probleme_show', ['id' => $probleme->getId()]);
-        }
-
-        return $this->render($render, [
-            'probleme' => $probleme,
-            'adresse' => $probleme->getLocalisation(),
-            'form' => $form->createView(),
-            'images' =>$images
-        ]);
     }
 
     /**
@@ -292,16 +304,21 @@ class ProblemeController extends AbstractController
         ProblemeService $problemeService
     ): Response
     {
-        if(!$this->permissionChecker->isUserGranted(['DELETE_OTHER_PROBLEME'])){
-            $this->addFlash('fail','Vous ne possedez pas les permissions necessaires pour supprimer ce problème');
-            return $this->redirectToRoute('home_index');
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["DELETE_OTHER_PROBLEME"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            } else {
+                if ($this->isCsrfTokenValid('delete' . $probleme->getId(), $request->request->get('_token'))) {
+                    $problemeService->DeleteProbleme($probleme);
+                }
+                $this->addFlash('success', 'le problème "' . $probleme->getTitre() . '" a bien été supprimé');
+                return $this->redirectToRoute('probleme_index');
+            }
         }
-
-        if ($this->isCsrfTokenValid('delete'.$probleme->getId(), $request->request->get('_token'))) {
-           $problemeService->DeleteProbleme($probleme);
-        }
-        $this->addFlash('success','le problème "'.$probleme->getTitre().'" a bien été supprimé');
-        return $this->redirectToRoute('probleme_index');
     }
 
     /**
@@ -317,8 +334,6 @@ class ProblemeController extends AbstractController
         PrioriteRepository $prioriteRepository
     ): Response
     {
-
-
         $form = $this->createForm(RedirectProblemeType::class);
         $form->handleRequest($request);
         $probleme = new Probleme();
@@ -368,25 +383,29 @@ class ProblemeController extends AbstractController
         StatutRepository $statutRepository
     ) : Response
     {
-        if(!$this->permissionChecker->isUserGranted(['VALIDATE_PROBLEME'])){
-            $this->addFlash('fail','Vous ne possedez pas les permissions necessaires pour valider ce problème');
-            return $this->redirectToRoute('home_index');
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["EDIT_OTHER_HISTORIQUE_STATUT"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            } else {
+                if ($probleme->getHistoriqueStatuts()->last()->getStatut()->getNom() != "Nouveau") {
+                    return $this->redirectToRoute('probleme_show', ["id" => $probleme->getId()]);
+                }
+
+                $entityManager = $this->getDoctrine()->getManager();
+
+                $hs = new HistoriqueStatut();
+                $hs->setProbleme($probleme);
+                $hs->setStatut($statutRepository->findOneBy(["nom" => "Ouvert"]));
+                $hs->setDate(new \DateTime());
+
+                $entityManager->persist($hs);
+                $entityManager->flush();
+            }
         }
-        if($probleme->getHistoriqueStatuts()->last()->getStatut()->getNom() != "Nouveau"){
-            return $this->redirectToRoute('probleme_show', ["id" => $probleme->getId()]);
-        }
-
-        $entityManager = $this->getDoctrine()->getManager();
-
-        $hs = new HistoriqueStatut();
-        $hs->setProbleme($probleme);
-        $hs->setStatut($statutRepository->findOneBy(["nom" => "Ouvert"]));
-        $hs->setDate(new \DateTime());
-
-        $entityManager->persist($hs);
-        $entityManager->flush();
-
-
         return $this->redirectToRoute('probleme_show', ["id" => $probleme->getId()]);
     }
 
@@ -398,24 +417,29 @@ class ProblemeController extends AbstractController
         StatutRepository $statutRepository
     ) : Response
     {
-        if(!$this->permissionChecker->isUserGranted(['UPDATE_OTHER_HISTORIQUE_STATUT'])){
-            $this->addFlash('fail','Vous ne possedez pas les permissions necessaires pour archiver ce problème');
-            return $this->redirectToRoute('home_index');
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["UPDATE_OTHER_HISTORIQUE_STATUT"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            } else {
+                if ($probleme->getHistoriqueStatuts()->last()->getStatut()->getNom() != "Résolu") {
+                    return $this->redirectToRoute('probleme_show', ["id" => $probleme->getId()]);
+                }
+
+                $entityManager = $this->getDoctrine()->getManager();
+
+                $hs = new HistoriqueStatut();
+                $hs->setProbleme($probleme);
+                $hs->setStatut($statutRepository->findOneBy(["nom" => "Archivé"]));
+                $hs->setDate(new \DateTime());
+
+                $entityManager->persist($hs);
+                $entityManager->flush();
+            }
         }
-        if($probleme->getHistoriqueStatuts()->last()->getStatut()->getNom() != "Résolu"){
-            return $this->redirectToRoute('probleme_show', ["id" => $probleme->getId()]);
-        }
-
-        $entityManager = $this->getDoctrine()->getManager();
-
-        $hs = new HistoriqueStatut();
-        $hs->setProbleme($probleme);
-        $hs->setStatut($statutRepository->findOneBy(["nom" => "Archivé"]));
-        $hs->setDate(new \DateTime());
-
-        $entityManager->persist($hs);
-        $entityManager->flush();
-
 
         return $this->redirectToRoute('probleme_show', ["id" => $probleme->getId()]);
     }
@@ -428,20 +452,28 @@ class ProblemeController extends AbstractController
         StatutRepository $statutRepository
     ) : Response
     {
-        if($probleme->getHistoriqueStatuts()->last()->getStatut()->getNom() != "Archivé"){
-            return $this->redirectToRoute('probleme_show', ["id" => $probleme->getId()]);
+        if(!$this->isGranted('ROLE_USER')){
+            $this->addFlash('fail','Veuillez vous connectez pour acceder à cette page.');
+            return $this->redirectToRoute('app_login');
+        }else {
+            if (!$this->permissionChecker->isUserGranted(["UPDATE_OTHER_HISTORIQUE_STATUT"])) {
+                $this->addFlash('fail', 'Vous ne possedez pas les permissions necessaires.');
+                return new RedirectResponse("/");
+            } else {
+                if ($probleme->getHistoriqueStatuts()->last()->getStatut()->getNom() != "Archivé") {
+                    return $this->redirectToRoute('probleme_show', ["id" => $probleme->getId()]);
+                }
+                $entityManager = $this->getDoctrine()->getManager();
+
+                $hs = new HistoriqueStatut();
+                $hs->setProbleme($probleme);
+                $hs->setStatut($statutRepository->findOneBy(["nom" => "Ouvert"]));
+                $hs->setDate(new \DateTime());
+
+                $entityManager->persist($hs);
+                $entityManager->flush();
+            }
         }
-        $entityManager = $this->getDoctrine()->getManager();
-
-        $hs = new HistoriqueStatut();
-        $hs->setProbleme($probleme);
-        $hs->setStatut($statutRepository->findOneBy(["nom" => "Ouvert"]));
-        $hs->setDate(new \DateTime());
-
-        $entityManager->persist($hs);
-        $entityManager->flush();
-
-
         return $this->redirectToRoute('probleme_show', ["id" => $probleme->getId()]);
     }
